refactor into subdirectories
organise C++ files into directories called "device" (for hardware video input devices such as webcams), "opencv" (for OpenCV computer vision), "ui" (for cvui GUI using OpenCV)
