+++ video/device/macos_list_devices.h
