+++ video/device/win32_list_devices.cpp
