+++ video/opencv/exit_requested.cpp
