+++ video/opencv/exit_requested.h
