+++ video/opencv/take_picture.cpp
