+++ video/ui/main_window.h
@@ -7,8 +7,8 @@
 // EnhancedWindow.h must be included after cvui.h
 #include "EnhancedWindow.h"
 
-#include "input_device.h"
-#include "opencv_window.h"
+#include "../device/input_device.h"
+#include "../opencv/opencv_window.h"
 
 namespace frank::video {
